# Improve file loading for .img files.