--- a/rwengine/src/engine/GameData.cpp
+++ b/rwengine/src/engine/GameData.cpp
@@ -41,7 +41,11 @@ GameData::GameData(Logger* log, const rwfs::path& path)
         });
 }
 
-void GameData::load() {
+bool GameData::load() {
+    if (!isValidGameDirectory()) {
+        return false;
+    }
+
     index.indexTree(datpath);
 
     loadIMG("models/gta3.img");
@@ -78,6 +82,8 @@ void GameData::load() {
 
     // Load ped groups after IDEs so they can resolve
     loadPedGroups("data/pedgrp.dat");
+
+    return true;
 }
 
 void GameData::loadLevelFile(const std::string& path) {
@@ -759,13 +765,11 @@ float GameData::getWaveHeightAt(const glm::vec3& ws) const {
            WATER_HEIGHT;
 }
 
-bool GameData::isValidGameDirectory(const rwfs::path& path) {
+bool GameData::isValidGameDirectory() const {
     rwfs::error_code ec;
-    if (!rwfs::is_directory(path, ec)) {
-        std::cerr << "first test failed\n";
+    if (!rwfs::is_directory(datpath, ec)) {
         return false;
     }
 
-    LoaderIMG i;
-    return i.load(path / "models/gta3.img");
+    return !ec;
 }