Hi again!

So while working on openrw/debugging an issue, I really started to notice the significant loading time until being in-game.

I did some perf'ing and noticed that `LoaderIMG::load` was a pretty hot path and after digging a bit, I found out that `LoaderIMG::load` was called 3237(!) times with `models/gta3.img`.
That meant opening and reading the file and allocating memory for it each call.

I did some simple optimization, now each file is only loaded once by `LoaderIMG`, which resulted in a pretty nifty speedup. 
Also after these changes, `LoaderIMG::load` completely vanishes from the flamegraph.

On a hot file cache, loading on startup takes ~7600ms on `master` on my machine, after the changes it goes down to ~5100ms.
There might be even more loading time reduction on slower systems and/or spinning rust.

Only tested on Linux, so I’d much appreciate if someone could test this on Windows to check if really nothing breaks.