@mole99 2020-05-16
I tested this PR on Linux and got a nice speedup:

| Master | PR |
|:-:|:-:|
| 3283 ms | 2509 ms |
| 3280 ms | 2460 ms |
| 3272 ms | 2500 ms |

That's a decrease of around 24% in loading time. Awesome!