@madebr 2020-05-19
@christoph-heiss 
These changes will make the `LoaderIMG` thread unsafe.
So maybe add a note at the top of the `LoaderIMG` class that the class is not thread safe and then this is good to merge.