@christoph-heiss 2020-05-19
> So maybe add a note at the top of the LoaderIMG class that the class is not thread safe and then this is good to merge.

Good suggestion, just pushed another commit which adds an appriopriate note to the class.

And if sometimes in the future the usecase arrives that we need to use this from multiple threads, making it MT-safe is fairly easy.