@madebr 2020-05-20
Thanks @christoph-heiss !

> Good suggestion, just pushed another commit which adds an appriopriate note to the class.

This suggestion was made by @danhedron on the irc channel.